Log truncated beacons implicated in WiFi issues
This change adds a log trace to track the reception of apparently truncated
WiFi Beacons and Probe Responses.

This is a follow up to issue #25:
"Squeezebox Radio - Some notes on WiFi stability problems"

I have noted that WiFi Beacons and Probe Responses are being truncated to 496
octets in length, probably by the Atheros firmware. Together with a 16 octet
packet header this equates to a packet size of 512 octets.

The process of truncation is, somehow, triggering the Radio's WiFi
misbehaviour.

By tracing these occurrences, we may gain some insight into:

- How many of these are being received in any given environment.
- How many will result in WiFi misbehaviour.
- Whether time of receipt is relevant to occurrence of misbehaviour.