WiFi Robustness Applet - Bug fixes
This change gives effect to the following:

Triggers WiFi restart when the user exits the menu, rather than simply
navigating to another window. This was not an issue until some recent changes
added some sub-menus to the Robustness Applet's main menu. Navigating to these
sub-menus would trigger a WiFi restart even though the user has not completed
the task.

Ensures that a wlan.conf file is generated even if it has been inadvertently
deleted. At present the function '_fileSub' will simply fail if the file
cannot be opened for reading. This is most likely to occur if the file has
been inadvertently deleted. This change allows the function to proceed and
write out an updated wlan.conf.

Tidies up the calls to wmiconfig in the wlan start up script. The correct
form of these commands should be, for example, wmiconfig -i eth1 --wmode gonly.